Hold the CI token to read, and tidy two allocations
The workflow files took whatever the repository's default token
permissions happened to be. These jobs only read the repository in order
to build it, so say so.

Two unrelated tidy-ups found while reading:

  - term_win_init() multiplied two ints before widening to size_t, so a
    large enough terminal would have overflowed the allocation size.
    Cast first.
  - lookup_symbol() copied a bracketed name into a temporary buffer just
    to trim two characters off it.  A precision specifier does the same
    thing without the allocation.

And a note to myself in Idea.md on how a release is cut.

+++ src/ui-term.c
@@ -361,16 +361,16 @@ static errr term_win_init(term_win *s, int w, int h)
 	s->c = mem_zalloc_alt(h * sizeof(wchar_t*));
 
 	/* Make the window content arrays */
-	s->va = mem_zalloc_alt(h * w * sizeof(int));
-	s->vc = mem_zalloc_alt(h * w * sizeof(wchar_t));
+	s->va = mem_zalloc_alt((size_t)h * w * sizeof(int));
+	s->vc = mem_zalloc_alt((size_t)h * w * sizeof(wchar_t));
 
 	/* Make the terrain access arrays */
 	s->ta = mem_zalloc_alt(h * sizeof(int*));
 	s->tc = mem_zalloc_alt(h * sizeof(wchar_t*));
 
 	/* Make the terrain content arrays */
-	s->vta = mem_zalloc_alt(h * w * sizeof(int));
-	s->vtc = mem_zalloc_alt(h * w * sizeof(wchar_t));
+	s->vta = mem_zalloc_alt((size_t)h * w * sizeof(int));
+	s->vtc = mem_zalloc_alt((size_t)h * w * sizeof(wchar_t));
 
 	/* Prepare the window access arrays */
 	for (y = 0; y < h; y++) {
